home/raffaello/spacewx_web/requirements.txt
Flask==3.0.3
pandas==2.2.2
python-dateutil==2.9.0.post0

